+++ NEWS.md
@@ -4,6 +4,7 @@
 - `transition()` now supports a named `list` model, e.g: list(yes = 0.1, no = 0.9).
 - Introduced `add_log()`, `get_log()`, `mutate_entity()`, and `add_entity()`. These functions are supposed to make constructing a microsimulation pipeline simpler. See their documentation for details.
 - Made `$colnames()` into an active field for `DataBackendDataFrame`.
+- Added `omit_derived_varnames()` for internal use.
 
 ## Minor changes